requirements.txt
langchain-groq
langchain
python-dotenv
pypdf
pandas
groq
